+++ Electron/src/components/AdvancedUIComponents/ContextMenu/Playlist/ContextMenuPlaylist.tsx
@@ -2,9 +2,9 @@ import Popover from '@mui/material/Popover';
 import { useCallback, useEffect, useReducer, useState } from 'react';
 import Global from 'global/global';
 import { useNavigate } from 'react-router-dom';
-import LoadingCircle from 'componentes/AdvancedUIComponents/LoadingCircle/LoadingCircle';
-import InfoPopover from 'componentes/AdvancedUIComponents/InfoPopOver/InfoPopover';
-import { InfoPopoverType } from 'componentes/AdvancedUIComponents/InfoPopOver/types/InfoPopover';
+import LoadingCircle from 'components/AdvancedUIComponents/LoadingCircle/LoadingCircle';
+import InfoPopover from 'components/AdvancedUIComponents/InfoPopOver/InfoPopover';
+import { InfoPopoverType } from 'components/AdvancedUIComponents/InfoPopOver/types/InfoPopover';
 import Token from 'utils/token';
 import { backendPathFromUserType } from 'utils/role';
 import styles from '../contextMenu.module.css';
