+++ Electron/src/components/AdvancedUIComponents/InfoPopOver/types/InfoPopover.ts
