+++ Electron/src/components/AdvancedUIComponents/LoadingCircle/LoadingCircle.tsx
