+++ Electron/src/components/Cards/ArtistCard/ArtistCard.tsx
