+++ Electron/src/components/Cards/ArtistCard/types/propsArtistCard.ts
