+++ Electron/src/components/Cards/GenreCard/GenreCard.module.css
