+++ Electron/src/components/Cards/GenreCard/GenreCard.tsx
