+++ Electron/src/components/Cards/PlaylistCard/PlaylistCard.tsx
@@ -1,6 +1,6 @@
 import { Link, useNavigate } from 'react-router-dom';
 import { useState, MouseEvent, useEffect } from 'react';
-import ContextMenuPlaylist from 'componentes/AdvancedUIComponents/ContextMenu/Playlist/ContextMenuPlaylist';
+import ContextMenuPlaylist from 'components/AdvancedUIComponents/ContextMenu/Playlist/ContextMenuPlaylist';
 import Popover, { PopoverPosition } from '@mui/material/Popover';
 import styles from '../cards.module.css';
 import { PropsPlaylistCard } from './types/propsPlaylistCard';
