+++ Electron/src/components/Cards/PlaylistCard/types/propsPlaylistCard.ts
