+++ Electron/src/components/Cards/SongCard/SongCard.tsx
@@ -1,23 +1,34 @@
-import { useState, useEffect, MouseEvent } from 'react';
-import { PropsSongs } from 'componentes/Sidebar/types/propsSongs.module';
-import ContextMenuSong from 'componentes/AdvancedUIComponents/ContextMenu/Song/ContextMenuSong';
+import { useEffect, useState } from 'react';
+import ContextMenuSong from 'components/AdvancedUIComponents/ContextMenu/Song/ContextMenuSong';
 import Popover, { PopoverPosition } from '@mui/material/Popover';
 import { useNavigate } from 'react-router-dom';
-import { secondsToMinutesSeconds } from 'utils/date';
-import styles from '../../pages/Playlist/playlist.module.css';
+import styles from '../cards.module.css';
+import defaultThumbnailPlaylist from '../../../assets/imgs/DefaultThumbnailPlaylist.jpg';
 
-export default function Song({
+export interface PropsSongCard {
+  name: string;
+  artist: string;
+  photo: string;
+  refreshSidebarData: Function;
+  changeSongName: Function;
+}
+
+export default function SongCard({
   name,
-  playlistName,
-  artistName,
-  index,
-  duration,
-  playCount,
-  handleSongCliked,
-  refreshPlaylistData,
+  artist,
+  photo,
   refreshSidebarData,
-}: PropsSongs) {
+  changeSongName,
+}: PropsSongCard) {
   const navigate = useNavigate();
+
+  const handleClickArtist = () => {
+    navigate(`/artist/${artist}`);
+  };
+
+  const handleClickSong = () => {
+    changeSongName(name);
+  };
 
   const [isOpen, setIsOpen] = useState(false);
   const [anchorPosition, setAnchorPosition] = useState<{
@@ -35,15 +46,7 @@ export default function Song({
     }
   }, [isOpen]);
 
-  const handleSongClicked = () => {
-    handleSongCliked(name);
-  };
-
-  const handleClickArtist = () => {
-    navigate(`/artist/${artistName}`);
-  };
-
-  const handleRightClick = (event: MouseEvent<HTMLLIElement>) => {
+  const handleRightClick = (event: any) => {
     event.preventDefault();
     setIsOpen(!isOpen);
     setAnchorPosition({
@@ -56,36 +59,36 @@ export default function Song({
   const id = open ? 'parent-popover' : undefined;
 
   return (
-    <li
-      onDoubleClick={handleSongClicked}
-      className={`container-fluid ${styles.gridContainer} align-items-center`}
-      onContextMenu={handleRightClick}
-    >
-      <span className={` ${styles.songNumberTable}`}>{index}</span>
-      <span className={`  d-flex flex-column`}>
-        <span
-          className={`${styles.songTitleTable} ${styles.titleContainer} pb-0`}
-        >
-          {name}
-        </span>
-
-        <div>
+    <>
+      <div
+        style={{
+          border: 'none',
+          backgroundColor: 'transparent',
+          textAlign: 'left',
+        }}
+        key={name + artist}
+        className={`rounded ${styles.card}`}
+        onDoubleClick={handleClickSong}
+        onContextMenu={handleRightClick}
+      >
+        <div className={`${styles.imgContainer}`}>
+          <img
+            src={photo === '' ? defaultThumbnailPlaylist : photo}
+            className="card-img-top rounded"
+            alt="song thumbnail"
+          />
+        </div>
+        <div className={`${styles.cardBody}`}>
+          <h5 className={`${styles.tituloLista}`}>{name}</h5>
           <button
             type="button"
             onClick={handleClickArtist}
-            className={`${styles.gridItem} ${styles.artistNameContainer} p-0 `}
+            className={`${styles.autorLista}`}
           >
-            {artistName}
+            {artist}
           </button>
         </div>
-      </span>
-      <span className={` d-flex justify-content-end ${styles.gridItem}`}>
-        {playCount}
-      </span>
-      <span className={` d-flex justify-content-end ${styles.gridItem}`}>
-        {secondsToMinutesSeconds(duration)}
-      </span>
-
+      </div>
       <div>
         <Popover
           id={id}
@@ -112,13 +115,13 @@ export default function Song({
         >
           <ContextMenuSong
             songName={name}
-            playlistName={playlistName}
+            playlistName=""
             handleCloseParent={handleClose}
-            refreshPlaylistData={refreshPlaylistData}
+            refreshPlaylistData={() => {}}
             refreshSidebarData={refreshSidebarData}
           />
         </Popover>
       </div>
-    </li>
+    </>
   );
 }
