+++ Electron/src/components/ShowAllItems/Items/ItemsAllPlaylists.tsx
@@ -1,35 +1,17 @@
 import { useCallback, useEffect, useState } from 'react';
-import { PropsPlaylistCard } from 'componentes/Cards/PlaylistCard/types/propsPlaylistCard';
-import PlaylistCard from 'componentes/Cards/PlaylistCard/PlaylistCard';
+import { PropsPlaylistCard } from 'components/Cards/PlaylistCard/types/propsPlaylistCard';
+import PlaylistCard from 'components/Cards/PlaylistCard/PlaylistCard';
 import Global from 'global/global';
-import { backendPathFromUserType } from 'utils/role';
-import { PropsItemsPlaylistsFromUser } from '../types/PropsItems';
+import { PropsItemsPlaylist } from '../types/PropsItems';
 import defaultThumbnailPlaylist from '../../../assets/imgs/DefaultThumbnailPlaylist.jpg';
 
-export default function ItemsAllPlaylistsFromUser({
+export default function ItemsAllPlaylists({
   refreshSidebarData,
-  userName,
-  userType,
-}: PropsItemsPlaylistsFromUser) {
+}: PropsItemsPlaylist) {
   const [playlists, setPlaylists] = useState<PropsPlaylistCard[]>();
 
-  const handlePlaylists = useCallback(async () => {
-    const playlistNames: string[] = [];
-
-    try {
-      const fetchUrlPlaylistFromUser = `${Global.backendBaseUrl}${backendPathFromUserType[userType]}/${userName}`;
-
-      const resFetchUrlPlaylistFromUser = await fetch(fetchUrlPlaylistFromUser);
-      const resFetchUrlPlaylistFromUserJson =
-        await resFetchUrlPlaylistFromUser.json();
-
-      playlistNames.push(resFetchUrlPlaylistFromUserJson.playlists.join(','));
-    } catch {
-      console.log('Unable to get user data');
-      return;
-    }
-
-    fetch(`${Global.backendBaseUrl}playlists/multiple/${playlistNames}`)
+  const handlePlaylists = useCallback(() => {
+    fetch(`${Global.backendBaseUrl}playlists/`)
       .then((resFetchPlaylists) => resFetchPlaylists.json())
       .then((resFetchPlaylistsJson) => {
         if (resFetchPlaylistsJson.playlists) {
@@ -59,7 +41,7 @@ export default function ItemsAllPlaylistsFromUser({
       .catch(() => {
         console.log('No se pudieron obtener las playlists');
       });
-  }, [refreshSidebarData, userName, userType]);
+  }, [refreshSidebarData]);
 
   useEffect(() => {
     handlePlaylists();
