+++ Electron/src/components/ShowAllItems/Items/ItemsAllSongsFromArtist.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
-import SongCard, { PropsSongCard } from 'componentes/Cards/SongCard/SongCard';
+import SongCard, { PropsSongCard } from 'components/Cards/SongCard/SongCard';
 import Global from 'global/global';
 import { PropsItemsSongsFromArtist } from '../types/PropsItems';
