+++ Electron/src/components/ShowAllItems/ShowAllItems.tsx
