+++ Electron/src/components/ShowAllItems/types/PropsItems.ts
