+++ Electron/src/components/ShowAllItems/types/PropsShowAllItems.ts
