+++ Electron/src/components/Sidebar/ModalAddSongPlaylist/Accordion/AddSongPlayListAccordion.tsx
@@ -8,10 +8,10 @@ import LibraryMusicRoundedIcon from '@mui/icons-material/LibraryMusicRounded';
 import AudiotrackIcon from '@mui/icons-material/Audiotrack';
 import Global from 'global/global';
 import Token from 'utils/token';
-import { InfoPopoverType } from 'componentes/AdvancedUIComponents/InfoPopOver/types/InfoPopover';
-import ConfirmationModal from 'componentes/AdvancedUIComponents/InfoPopOver/InfoPopover';
+import { InfoPopoverType } from 'components/AdvancedUIComponents/InfoPopOver/types/InfoPopover';
+import ConfirmationModal from 'components/AdvancedUIComponents/InfoPopOver/InfoPopover';
 import { UserType } from 'utils/role';
-import LoadingCircleSmall from 'componentes/AdvancedUIComponents/LoadingCircle/LoadingCircleSmallNoPadding';
+import LoadingCircleSmall from 'components/AdvancedUIComponents/LoadingCircle/LoadingCircleSmallNoPadding';
 import GenreOption from './GenreOption/GenreOption';
 import styles from './addSongPlayListAccordion.module.css';
