+++ Electron/src/components/Sidebar/ModalAddSongPlaylist/Accordion/GenreOption/GenreOption.tsx
