+++ Electron/src/components/Sidebar/Sidebar.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState, useCallback } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import Global from 'global/global';
 import Token from 'utils/token';
-import LoadingCircle from 'componentes/AdvancedUIComponents/LoadingCircle/LoadingCircle';
+import LoadingCircle from 'components/AdvancedUIComponents/LoadingCircle/LoadingCircle';
 import { backendPathFromUserType } from 'utils/role';
 import styles from './sideBarCss.module.css';
 import Playlist from './Playlist/Playlist';
