+++ Electron/src/components/Sidebar/types/propsPlaylist.module.ts
