+++ Electron/src/components/footer/Footer.tsx
