+++ Electron/src/components/footer/Player/Player.tsx
