+++ Electron/src/components/footer/Player/TimeSlider/TimeSlider.tsx
