+++ Electron/src/components/footer/Player/TimeSlider/timeSlider.module.css
