+++ Electron/src/components/footer/SongConfig/SongConfig.tsx
