+++ Electron/src/components/footer/SongConfig/VolumeSlider/VolumeSlider.tsx
