+++ Electron/src/components/footer/SongConfig/VolumeSlider/volumeSlider.module.css
