+++ Electron/src/components/footer/SongConfig/songConfig.module.css
