+++ Electron/src/components/footer/SongInfo/SongInfo.tsx
