+++ Electron/src/components/footer/SongInfo/songInfo.module.css
